chore(text-field): move to packages
