+++ packages/components/fields/text-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './text-field';
+
+export { default as version } from './version';
