+++ packages/components/fields/text-field/src/text-field.form.story.js
@@ -6,11 +6,11 @@ import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, select } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
-import Section from '../../../../.storybook/decorators/section';
-import FormikBox from '../../../../.storybook/decorators/formik-box';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import FormikBox from '../../../../../.storybook/decorators/formik-box';
+import Readme from '../README.md';
 import TextField from './text-field';
-import TextInput from '../../inputs/text-input';
+import TextInput from '../../../inputs/text-input';
 
 // Cool stuff to try in this story:
 //  - Click the "Username" label and see how the input is focused automatically
