+++ packages/components/fields/text-field/src/text-field.story.js
@@ -9,9 +9,9 @@ import {
   select,
   object,
 } from '@storybook/addon-knobs/react';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
-import * as icons from '../../icons';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
+import * as icons from '../../../icons';
 import TextField from './text-field';
 
 storiesOf('Components|Fields', module)
