+++ packages/components/fields/text-field/src/text-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TextField } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = 'hello world, how are you?';
