--- src/components/fields/text-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './text-field';
